Move some functions around into different files.
