+++ lib/src/al/env.c
@@ -24,7 +24,7 @@
 #include "types.h"
 #include "macros.h"
 #include "audio_internal.h"
-#include "thread0_epc.h"
+#include "unknown_B7B30.h"
 
 // TODO: these come from headers
 #ident "$Revision: 1.17 $"
@@ -111,7 +111,6 @@ Acmd *alEnvmixerPull(void *filter, UNUSED s16 *outp, s32 outCount, s32 sampleOff
         assert(samples <= AL_MAX_RSP_SAMPLES);
 #else
         // Something must have gone wrong when compiling this file, and the asserts got left in.
-        // The function happens to currently live in the thread0 file for some reason.
         if (samples >= 0) {} 
         else { thread0_Assert("samples >= 0", "env.c", 104); }
         if (samples <= AL_MAX_RSP_SAMPLES) {}
