+++ src/thread0_epc.c
@@ -1,5 +1,5 @@
 /* The comment below is needed for this file to be picked up by generate_ld */
-/* RAM_POS: 0x800B6E50 */
+/* RAM_POS: 0x800B6F50 */
 
 #include "thread0_epc.h"
 #include "types.h"
@@ -41,43 +41,6 @@ extern OSMesgQueue D_801297E8;
 extern OSMesg D_801297C8;
 extern u64 gEPCStack[0x200];
 extern s32 D_80129FB0[3];
-
-void func_800B6E50(void) {
-    if (D_80127CD0 < 0x141) {
-        D_80127CBC = 0x10;
-        D_80127CC0 = D_80127CD0 - 0x10;
-    } else {
-        D_80127CBC = 0x20;
-        D_80127CC0 = D_80127CD0 - 0x20;
-    }
-    if (D_80127CD2 < 0xF1) {
-        D_80127CC4 = 0x10;
-        D_80127CC8 = D_80127CD2 - 0x10;
-    } else {
-        D_80127CC4 = 0x20;
-        D_80127CC8 = D_80127CD2 - 0x20;
-    }
-}
-
-void func_800B6EE0(void) {
-    D_80127CAC = D_80127CBC;
-    D_80127CAE = D_80127CC4;
-}
-
-void func_800B6F04(void) {
-    D_80127CAC = D_80127CBC;
-    D_80127CAE += 11;
-}
-
-void func_800B6F30(UNUSED s32 arg0, UNUSED s32 arg1, UNUSED s32 arg2) {}
-
-// Only called from two libultra funcs for some reason. 
-// This is likely an assert gone wrong.
-// Called from alEnvMixerPull, and _pullSubFrame
-// thread0_Assert("samples >= 0", "env.c", 104);
-// thread0_Assert("samples <= AL_MAX_RSP_SAMPLES", "env.c", 105);
-// thread0_Assert("source", "env.c", 373);
-void thread0_Assert(UNUSED char *message, UNUSED char *fileName, UNUSED s32 lineNumber) {}
 
 /**
  * Start the exception program counter thread.
