+++ src/unknown_B7B30.h
@@ -1,0 +1,10 @@
+#ifndef _UNKNOWN_B7B30_H_
+#define _UNKNOWN_B7B30_H_
+
+#include "types.h"
+#include "libultra_internal.h"
+#include "macros.h"
+
+void thread0_Assert(UNUSED char *message, UNUSED char *fileName, UNUSED s32 lineNumber);
+
+#endif
